1.fix init_param computing bug( s2e2 with goal basis) 2. devide tau before returning vel and acc

+++ mp_pytorch/basis_gn/uni_bspline_basis.py
@@ -11,7 +11,7 @@ class UniBSplineBasis(BasisGenerator):
                  phase_generator: LinearPhaseGenerator,
                  num_basis: int = 10,
                  degree_p: int = 3,
-                 dtype: torch.dtype = torch.float64,
+                 dtype: torch.dtype = torch.float32,
                  device: torch.device = 'cpu',
                  **kwargs):
         """
